requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
APScheduler==3.10.4
email-validator==2.1.0
python-dotenv==1.0.0
werkzeug==3.0.1
anthropic==0.39.0
libsql-experimental==0.0.55
sqlalchemy-libsql==0.2.0
